new/src/collector/requirements.txt
numpy==1.19.4
websocket_client==0.57.0
